requirements.txt
faiss-cpu==1.13.0
fastapi==0.120.1
keras==3.12.0
langchain==1.0.2
langchain-community==0.4.1
langchain_google_genai==3.1.0
langchain-openai==1.0.1
langchain-text-splitters==1.0.0
mangum==0.19.0
nltk==3.9.2
numpy==2.3.4
python-dotenv==1.2.1
sentence-transformers==5.1.2
tensorflow==2.20.0
tf-keras==2.20.1
uvicorn==0.38.0